% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentist.R
\name{print.dentist}
\alias{print.dentist}
\title{Print dentist
print summary of output from dent_walk}
\usage{
\method{print}{dentist}(dentist_object, ...)
}
\arguments{
\item{dentist_object}{An object of class dentist}

\item{...}{Other arguments (not used)}
}
\description{
Print dentist
print summary of output from dent_walk
}
