% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentist.R
\name{sphere_propose}
\alias{sphere_propose}
\title{Propose new values based on points near bounds
This proposes new values based on wiggling a little bit away from the values near the bounds of the good region}
\usage{
sphere_propose(
  results,
  delta,
  old_params,
  lower_bound = -Inf,
  upper_bound = Inf,
  sd = 1
)
}
\arguments{
\item{results}{Data.frame of results so far}

\item{delta}{The maximum distance from the best value to consider}

\item{old_params}{The original parameter values}

\item{lower_bound}{Minimum parameter values to try. One for all or a vector of the length of par.}

\item{upper_bound}{Maximum parameter values to try. One for all or a vector of the length of par.}

\item{sd}{Standard deviation to use for the proposals. One for all or a vector of the length of par.}
}
\value{
A vector of the new parameter values
}
\description{
Propose new values based on points near bounds
This proposes new values based on wiggling a little bit away from the values near the bounds of the good region
}
